organised code, implemented global state
