+++ frontend/src/hooks/useLogin.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { useAuthenticationContext } from "../context/AuthenticationContext";
+import { useAuthenticationContext } from "./useAuthenticationContext";
 
 /* Custom Login Hook */
 export const useLogin = () => {
